Makefile: Prettify output and clean up
The main thing is to prettify the output, Whereas before you would get
output like:

    gcc -Wall -std=c99 -O2 -g -Wp,-D_FORTIFY_SOURCE=2 -D_FILE_OFFSET_BITS=64
    -fexceptions -fstack-protector --param=ssp-buffer-size=4 -fPIE -c
    utils.c -I../../libctemplate `pkg-config --cflags glib-2.0` `pkg-config
    --cflags gmime-2.0` `mysql_config --cflags`

for each C file. We now take a leaf from the Kernel and produce output
like:

      CC	compositus.o

This grealty reduces the clutter and makes compiler warnings easier to
see.

Also set an objects variable that contains all the .o files to negate
the need to list them all out several times.

And remove the url_handlers.h file from the url_handlers.o target as
that header is not included there.

Signed-off-by: Andrew Clayton <andrew@opentechlabs.co.uk>